refactor: move controllers to separate folder
